+++ Changelog.md
@@ -38,3 +38,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix broken "No entry" sign
+
+## [0.4.1] - Fri Jan 19 09:47:36 PST 2024
+
+### Fixed
+
+- Exception handling for curses error in footer
+
+## [0.4.2] - Fri Jan 19 09:47:36 PST 2024
+
+### Changed
+
+- Add year to date display
+- Right align "Blocked" icon
+
+## [0.5.0] - Thu Feb 29 19:06:13 PST 2024
+
+### Added
+
+- Script for CRON to block IP's with attack numbers > 1,000